sysdb: Handle extra args during instantiation
In entity.go I changed the map of attributes passed to InstantiateChild
to be the attributes map that was passed to createChildWithDef instead
of all the attributes of the entity. This allowed me to efficiently know
which attributes were set and which were there default value.

In places I needed to recover the attributes of the entity, to do that I
had to iterate over the typedef's attributes and call GetAttribute on
the entity for each one. Which is less efficient than having the
child.attributes map.

Change-Id: I37d6c5f9f1e875309ccd545d64ec876abf1f1b78