Fixing bug in logic to find test includes in playbooks
Change-Id: Ia70e42036885cd06d622c620ae972b8a463c2c0e

Temporarily removing all testing to allow for Ansible
package upgrade. Ansible 2.0 RPM was removed from Fedora
archives and replaced with Ansible 2.1 RPM. To avoid
further unnecessary changes in the future due to
dependency on Fedora archives, downloaded a local copy of
Ansible 2.1 RPM, with which Ansible would be installed
on the test Docker containers now.

This was to address the bug of all all testing for commits
pushed to ardc-config failing due to Ansible not being
properly installed, due to the missing RPM in the archives.

Change-Id: I69877ca608df38f39338c0e8ecea7dbca952159c