Checking in code for automatic ticket issuing/authentication.
This is set of files/code required from ASB to enable the automation.

The files committed under /scripts/auth/ denote the files
that are on ticketserver.
asbAuthenticator: cron job that runs every minute
asbAuthenticator.py: checks for server placeholder file that acts as
     whitelist. If placeholder file exists for a server, call
     restore_server script (twieland@) that reauthenticates P4
     tickets for that server.
restore_server: Thomas' (twieland@) script. Checking it in here so
     that we commit it to a code base.

auth step in ASB is the final stage.

Minor change in ASB-trigger, caused due to Gerrit URL change.

Change-Id: I002fb55c2bb5355b663750b21758b8e963048a52