entity, provider, smash, sysdb: Makes some notif timestamps consistent.
When we create or delete child entities, there were cases where we emit
more than one notification (particularly when we change a collection and
need to craft an extra "_counts" notification) and all the notifications
created as part of the same operation should have the same timestamp, but
the InstantiateChild() / DeleteChild() APIs didn't make it possible to
pick the timestamp upfront so it would be the same in all the notifs.

This change fixes the API and the code to enable just that.

Change-Id: I25aa68388ad274c46d571ecdd3ddc48e3764e7c1