entity, provider, smash, sysdb: Report collection `_counts` atomically.
When we update a collection in "streaming mode" (i.e. we're streaming
state to a backend, as oppose to being a more regular EOS agent like
OpenConfig) we emit updates for a synthesized "_counts" collection so
that the backend can keep track of (and subscribe to) the number of
entries in a collection without having to read the entire collection.

The "_counts" notification was created by taking the length of the
collection being updated outside of the critical section where the
collection is updated, leading to a potential race where by the time
we get the count, the collection has been updated again and the count
is thus off.

The fix is to read the count during the critical section -- to the
extent that it's possible (there are a couple TODOs in the code where
it wasn't easy to do this).

Two additional changes in Smash:
  - shadow.go: use the timestamp of the original notif instead of
    calling time.Now() again.
  - shadow_test.go: overwrite the timestamp of the expected notif
    instead of the actual notif to avoid a silly test-only data race
    (we don't care what the timestamps are anyways during these tests,
    which is why we overwrite it).

Change-Id: I2ffb8978ce5cd38935f553519e0dc2038a4df6c5