provider: Make the loopback provider easier to use.
When using the loopback provider with Agent (as opposed to on its own,
as is often the case in simpler tests) it's a bit hard to set things
up properly because the loopback provider needs to be initialized with
the producer chan from the Agent, and the agent needs to be initialized
with the loopback provider.  Chicken and egg problem.

So instead when one intends to use the loopback provider with an Agent,
one can just do NewLoopback(nil) and then the loopback provider will use
the agent's producer chan automatically.

Change-Id: I848513f13396dd5c836be0790ec8d0fa475e1090