added support for rolling restart
restarts scylla-server on one node at a time
after restart the server health is checked before proceeding further
When errors are not detected then restart continues with next node in
sequence, but if any errors are detected then there is an option to Continue
or Abort the rolling restart. Its quite possible that certain errors are
not really errors and so such prompt give the admin a chance to review
the log messages.

patchset-2: whitespace fix

Change-Id: I953364b5017bb9cb9d4ffe8f509fae959a6bb088