ops/ansible: Move to multiple inventories
Right now we're using one inventory and we're using
variables to define the different roles of each host.
This is not really the ansible way to do this as we're not
applying roles to hosts directly in the inventory.
In order to change this, we're moving from one big
inventory file to one inventory file per cluster.
This way we can just group the hosts by roles and not
use variables.

Change-Id: Ia814d806d49702af321079e97efe70026765f440