Ansible-scylla: Changed from yum module to shell
By default we have packages which meet "scylla*" disabled from ToolsV2
repo on all scylla servers. This means that when install a package viz
scylla-1.x.y we need to use the yum option of "--disableexcluded=all".
This option is not available in the ansible's yum module, hence switched
to shell.

I also tried "--disableexcluded=scylla-\*" and
"--disableexcluded=scylla-*" but neither of them worked.

Change-Id: I04f21ba1095c0a701fa56f1963b72998113deaa1