ops/ansible: Simplify task by using shared group_vars
the k8s_cluster role is installing some files on the hosts.
Those files are different for master and workers.
As masters and workers hosts are in two different groups
we create a playbook groupvars for each group listing
the files to deploy.
This way, we only have one task to install the files and
the correct ones will be picked up automatically based
on the group the host belongs to.

Change-Id: I4c0f831415f78520886adcba8c39b9aaa1e205f8