arista: Remove pointer for notification Updates
Based on a comment from Andrey Elenskiy, we don't need
to have updates as a pointer to a map.
IIRC, historically we did this so the jsonified
version of the notif will have nil when the updates
are empty.
Now we have a custom json marshaler which can take care
of this.
Removing the pointer might also remove the pointer for
each notification so the process should use less memory.

Change-Id: If7125b2d8db3cb2921484dff3e41d45e530e1f34