ops-docker: Use specific folder for docker.sock
The jenkins slaves are running inside containers and are
using the docker daemon from the host.
In order to access the docker daemon, the unix socket is
mounted in the container.
The problem is that when the daemon is restarted, it
recreates the socket and the container is loosing it
as it is still pointing to the old socket.
In order to fix this, the docker daemon is also creating
and docker socket in a specific folder and the folder
(and not the file only) will be mounted in the container.
This way, when docker daemon is restarted, the container
will see the new file.

Change-Id: I11d70c83ed90f8e10ae680eb7723fbf55878980f