ops-k8s: Change clusters domain
Having the same external and internal domain for k8s
is a problem.
If a service is exposed externally, the hostname of
this service will be <service>.<k8s_domain>
But if the k8s internal dns is set with the base domain
<k8s_domain> as well, the hostname "<service>.<k8s_domain>"
will not go to the external dns, but the k8s dns will try
to resolve it.
As internally, services hostnames are in the form:
<service>.<namespace>.svc.<k8s_domain>, the internal dns
will not be able to resolve <service>.<k8s_domain> and it
will return:

** server can't find kafka.dev.corp.arista.io: NXDOMAIN

In order to avoid this, we change the internal k8s domain to:

internal.<k8s_domain>

This way, services running internally, but exposed by an external
dns can still have their full external dns hostname resolved
correctly.

The problem first appeared with "npmregistry.dev.corp.arista.io"

Change-Id: I461ec59bf82f4d98e39981a13937b56b2efded13