Revert "ops/ansible: Switch to kubelet-wrapper"
This reverts commit 6ef2008e135be614c0345da26ebeecfdf3bde0ac.

(And does some changes in the way the binary is installed)

It looks like CoreOS does not plan to maintain this
kubelet-wrapper script in the long term.
Also, the images from quay.io are not updated with latest version.
There seem to be a delay (1.8 is still not there).
Using the binary gives us freedom to do whatever we want,
and be independent from the distro as well.

Change-Id: I50650ca60de1d18d2047df5b539f2b540dee25ce