ops-gcp: Add persistent storage for letsencrypt pod
Restarting the pod forces the start script to regenerate
the certificates. It accounts into the Let's Encrypt
rate limit.
In order to not ask for certificates we already have, a
persistent volume is added to the pod.
Now when the pod is restarted (To add new hostnames for instance)
the script can check if the cert is already there and not request
one.

This first commit is adding a helper pod used to initialize the
persistent storage. It might be used in the future to view/edit
the data in the persistent volume.

Change-Id: I215d98e44786bd9349b870316795580269133e25