ops/ansible: Add hostnames to /etc/hosts
We assume that we don't have external dns resolution.
Internally in the k8s cluster, all the services will talk
to each other using the k8s internal dns.
But externally, as we don't have a dns service, we
need to put the few external services into /etc/hosts
so we don't need dns setup.
Those external services are the following so far:
- etc servers (used by k8s master + haproxy service)
- k8s master (used by k8s workers kubelet)
- docker registry (used by k8s kubelet to pull the images to run)

Change-Id: Id9d9c3c10434b441c4e49fa4b63936c5e8d5c3f3