ops/ansible: Fix proxy config
Confd is using KUBERNETES_SERVICE_HOST to configure
its access to the apiserver. As this env var is set with
the apiserver ip address, the apiserver cert must have
this ip address defined.
As the internal k8s apiserver ip address can change, the external
ip address is used.
"kubernetes" service cannot be used here because the pod is using
hostNetwork: true which means the internal dns is not used.

This should be improved, probably by removing hostNetwork: true
and using "kubernetes" as hostname to access the apiserver.

Change-Id: I182f481a36af07a53e11cb6bc7aef54f126067cc