manage the error log review time
If errors are observed in the logs before or after an ops
activity then give the admin a chance to review the errors
before proceeding with the op activity.

Before starting the upgrade, we check errors but we don't need
to wait for 2mins. The 2min wait is needed only at the end of
the operation

Testing: tested the upgrade script on li101
Following is how the pause looks when errors are noticed

TASK [server_upgrader : Review the logs within next 1 hour and then
press Ctl+c to see further prompts] ***
Pausing for 3600 seconds
(ctrl+C then 'C' = continue early, ctrl+C then 'A' = abort)

patchset-2: Increased error log max pause time to 12hrs

Change-Id: Ia967b45151d893e3d5072b7e6f89a389c9cfe3c7