At request of navneet@ and twieland@ making sure AroraConfig-bs package is always available
Previous testing framework wasn't serving our purpose and was
failing all the builds so unforunately the change for that is also
part of this same commit.

The new testing framework copies over playboosk to be run onto
Ansible Master (ansible@) and from there runs on bs306 as the
test host. This most accurately simulates how our Ansible Push
framework is set up and run.

Change-Id: I5c2f12a8fcff19ad84a2050176d4cc3dca8eedf6