ops/ansible: Fix namenode init phase
In prod "kubectl logs" does not work.
As a result, the init phase using "command" and
"kubectl run --attach" is failing.
The workaround is to use the kubernetes ansible module
to start the pod.
As it is asynchronous, the user has to check the pod output
and the logs to validate the init phase has been correctly
executed.
As this is a one time execution it's ok to have this for now.
This init phase might be improved in the future.

Change-Id: Ie768880a4d462ea387f843bd534f7ac567b6ea5e