ops/ansible: Fix datanode config
The default value for dfs.datanode.failed.volumes.tolerated is
3 in the hardcoded config file.
This is ok for dev and staging as we have more than 3 volumes.
But on prod, we only have 3 volumes, so datanode does not want
to start with this current config value.
This change sets the config value to 3 when there are more than
three volumes mounted, and to 1 (value used for ovh-bhs) when
there is less than four volumes.

Related hadoop jira tickets:
https://issues.apache.org/jira/browse/HDFS-10269
https://issues.apache.org/jira/browse/HDFS-10279

Change-Id: I46bc5f16cb38c9d7cba92f52703b5ab1c36dd0d0