ops/ansible: Fix wrong etcd2 management
The etcd hostnames might not be defined in the dns.
In this case lookup(dig) won't work and this will result
in disabling etcd2 on all the nodes of the cluster,
effectivitely stopping etcd2 and blocking kubernetes.

This commit fixes the problem by using the ansible_host
value of the etcd hostname so the ip address is resolved
correctly.

Change-Id: I7dcfb23378bf70c4b8ac5f0315ccb85884c868e5