ops/ansible: Expose k8s services and domain
This change allows to access the services directly from
the laptop, including UIs.

Note that this change makes internal "arista.io" tools
like gerrit and jenkins unavailable.
so the user will have to disconnect from the vpn to access
the arista internal tools under "arista.io".

Example:

http://kafkaui.default.svc.internal.arista.io:9000/
http://hbasemaster-0.hbase.default.svc.internal.arista.io:16010/master-status

And links to region servers are working as well.

Change-Id: I0e62c60184b96050234e9c78a61907dfc6ce1e89