ops/ansible: Fix apiserver to kubelet host selection
By default the apiserver is using the node hostname to
connect to the kubelet.
The node hostname is resolving to the public ip by default.
And we bind the kubelet to the private ip.
So using kubectl logs|exec fails with "connection refused"
as apiserver cannot contact the kubelet.

This change tells apiserver to use the node ip first and not
the node name, os the correct private ip is used.

Change-Id: I4f12ddc24626ba94181e6e7a5f1a0b8724dbe411