ops/ansible: Add a letsencrypt domain variable
The current k8s_domain variable is used to define
the kubernetes internal domain.
It is also used to request letsencrypt certificates.
But for prod clusters, the letsencrypt domain is
different from the k8s internal domain, as it
is a public facing domain which will be used by the
customers to connect to Aeris Cloud.

This change is adding a letsencrypt_domain variable
defining this public domain so letsencrypt certificates
can be requested from the public domain.

Change-Id: I3c87d766e748a5b3da2951ccf5159cd276ef2465