types: Update Pointer to use path.Path
This change unfortunately has to be a little more complex
than just changing the previous types.Pointer in place to
have a path.Path instead of a string. The main reason for
this is that there are places where we are using pointers
as keys in maps and a pointer with an underlying path.Path
will cause this to fail (since a slice cannot be used as a
key in a map). We came up with two options to work around
this:

1) Add another hack to goarista/key that allows path.Path
   specifically to be used as a key in a map if it is in
   a key.Key. This was shown to be possible but has the
   consequence of causing an import cycle between the key
   and path package which could be difficult to resolve.
   See gerrit.corp.arista.io/c/33034.

2) Encode/decode the underlying path.Path to/from a string
   when the pointer needs to be used as a key in a map. A
   dose of added complexity comes along with this change,
   as well as decreased performance time for creating a
   pointer and accessing its path.

After some discussion, we ended up deciding on the second
option to avoid adding another hack to goarista/key and
having to deal with the import cycle there.

Change-Id: I366b72487cb1a366438086a168413b89bcab6280