provider, smash: pass collection path to NotificationsForCollectionCount
Calling entity.Path() is expensive as it will propogate up the tree.
As we are sending _count notification after actual update, we already
know the path and should reuse it.

This also fixes a bug in smash where a _count would be sent out for
path of shadow table's entity while update would be sent for path
of table's mount path. So if those differ, there would be an inconsistency.

Change-Id: Idce842814bf192fb746e24a55bfd7f03896e3768