adapter: Support plugins and add integration tests
Collector now supports a new flag -pluginDir, which is assumed to only
contain plugin files. If passed in, the adapter will load all plugins in
the directory. Plugins should register themselves in the exact same way
as normal built-in devices.

Testing is done using "make -C adapter", which once this commit is
merged should be run by a Jenkins job.

Change-Id: Ia2d6eb09b71d20bdff3ec65e89288618b39de841