cmd/Collector: Rework Adapter and Device configuration
We have been configuring the Adapter and Device with yaml files. There
are a few downsides to this: often there will only be a few options
actually used, and command-line flags would be easier; there's no great
way to define yaml config options and display options; and it requires
translating between yaml config and some of the configuration types we
already use in TerminAttr.

So all Adapter and Device configuration will be done with flags now:

Collector [adapter options] -device=<device> -deviceoption <key>=<val>

where the adapter options are fixed and defined in Collector/main.go and
deviceoption can be used repeatedly to set device option <key> to <val>.

Devices will now have to register a set of supported Options (i.e., pass
a map[string]Option to device.RegisterDevice), where Option is a struct
that gives the option's description, default value, and whether it's
required or not. Collector can use this to print a help string that
includes the Device's options. This has nothing to do with the Device
type in particular, and if it seems better to merge it with arista/flag,
that's doable too.

This also means that now Collector instantiates the Device directly
rather than passing it into the Adapter. This seems like a better
separation of concerns anyway, but it did mean a bunch of package
rearrangement was necessary, as now what adapter_test was doing should
now be in device_test, etc.

Change-Id: I0bf507bc4cd636b0056de2cdcaf5f4ebc65b885f