provider/snmp: Convert SNMP provider to use OpenConfig tree
I've added some library code that allows providers to use the OpenConfig
data tree defined in the gopenconfig package, and updated the SNMP
provider to use it. The library code sets up a data tree and feeds it
into a gNMI server that then produces gNMI notifications on relevant
updates. Right now we subscribe to all updates. gNMI notifications are
then converted into types.Notifications and streamed out as usual.

Using the OpenConfig tree gets us type checking and will force us to
adhere to the OpenConfig models. Using the gNMI server gets us
automatic notifications on changes and deletes.

The library code contains a number of APIs that the providers can use.
My approach has been to abstract away operations on the data tree as
much as possible. For now this is fine--having functions for updating
interface counters, LLDP data, and system data is tractable. Ultimately
we may want to take a more general approach, though it's not totally
clear how that would work. Touching the OpenConfig tree directly is a
lot of work and a bit finicky so some added structure is helpful.

Note: There's no delete functionality in the gNMI/OpenConfig library
yet because I haven't yet added it for SNMP.

Update: We're actually already doing the right thing as far as
generating path pointer notifications on leaf changes. So the Aeris
notification behavior should be correct.

Change-Id: I0967f54c5407542a76e1cc3d726ce2f1ec40c108