providers/snmp: Add more interface metrics to SNMP provider
I've added several more interface metrics to the SNMP provider and
cleaned up a bit inside the SNMP provider. Now we'll get all the
fields included in interface/state/counters except for in-fcs-errors,
carrier-transitions, and last-clear. We're also missing a few of the
leaves under interface/state: loopback-mode, description, enabled,
ifindex, and last-change. These will require more MIBs and I don't know
if it's worth it.

Change-Id: I5f81f5d2c88a5becd11e208957dcc5c7a72ae12d