arista: Refactor Provider interface
We are moving the original Provider interface as a generic interface
that captures the concept of a bare minimal provider, which allows up build
other provider interfaces that's not related to schema/entity/types.Notification.
As a result, we move the original provider.Provider interface to
provider.EOSProvider.

It also enables us to opensource our third-party device library :)

Doc:
https://docs.google.com/document/d/11W3El2g4GoCptVjqoqFLE1w1MEl-acfJbVth0pGMw74/edit#heading=h.9qpfzlsedz16

As we are removing Stop() and assume that if the context is
canceled it should be stopped, we need to wait for Run() to exit with a
nil error to make sure that it is indeed stopped and all resources are
cleaned up.

General refactor strategries:

In production code: Add a Setup function that stores the original arguments
from Run(). Change Run()’s prototype and all uses of original arguments to use
struct variables. If it has a meaningful Stop() before, Run() must be waiting
for some channel to be closed. Change that to be waiting for ctx.Done(), and calls
the original Stop() function which either does the cleanup itself or triggers
some other cleanups.

In tests: Change the original `go p.Run()` function call to:

```
c.Setup(schema, newDir(), nil)
ctx, cancel := context.WithCancel(context.Background())
var wg sync.WaitGroup
wg.Add(1)
go func() {
	err := c.Run(ctx)
	if err != nil {
	t.Error(err)
	wg.Done()
}()
```

And change the original `p.Stop()` function call to:

```
cancel()
wg.Wait()
```

Change-Id: I365e03043570263e32a6d809f97a9eca93d22dc4