provider: Use gNMI client in SNMP provider
This change was originally supposed to use the OpenConfig tree to
implement deletes for the SNMP provider, but ultimately that approach
was scrapped in favor of using the gNMI client instead of manipulating
the tree directly.

So I've added some infrastructure to the gnmi_helpers library to make
polling easier, and converted the SNMP provider to use it. The interface
is: OpenConfigPollForever(ctx, client, pollFn, interval, errChan).

OpenConfigPollForever calls the user-provided pollFn at the specified
interval. pollFn returns a gnmi.SetRequest, which OpenConfigPollForever
gives to the gNMI client to update the OpenConfig tree. Deletes are
implemented by including a Delete of the whole polled subtree (e.g.,
lldp or interface) and a set of Replace updates representing the current
state of the tree.

Change-Id: I6a129553c93739e671bda0c48b6237040c9eced5