provider: Convert SNMP and GNMI providers to GNMIProvider
I've updated the SNMP and GNMI providers to implement the GNMIProvider
interface rather than EOSProvider. I've also added an EOSProvider
implementation that takes in a GNMIProvider, which Adapter can use for
the conversion.

I've tested manually against a dut and CVP dut and the providers are
working fine. We should add unit testing but it'll probably have to wait
until we have some infrastructure for mock-testing.

Change-Id: Ia39769fa3e090c1a6478700c2e25b9b380277d07