provider: Eliminate gopenconfig dependency in SNMP provider
In preparation for moving the SNMP provider and related helper functions
into an open-source repo, I've changed the way the SNMP provider maps
from SNMP-defined type values to OpenConfig strings. We had been using
mappings defined in gopenconfig/eos and now we just define the mappings
in the provider package. OpenConfig actually uses the SNMP types, and
since we're trying to standardize to OpenConfig, they might plausibly be
used in the future. So I've put the mappings in
provider/openconfig/openconfig_helpers.go.

The hope is to switch from a provider/providers/<all providers>
directory structure  to provider/<provider name>/<provider_name>.go for
cleaner namespacing, etc., so I've started that here.

Change-Id: I44456866c2daeebaebac9216a70dc93086e87446