provider: Give each provider its own subpackage
I've refactored the provider package so that providers have their own
subpackages. This makes things less crazy, and since we may have quite a
lot of providers, it seems better to do it before we open-source it. So
whereas the structure been this:

arista/
    provider/
        eos_provider.go
        gnmi_provider.go
        provider.go
        providers/
            gnmi.go
            gnmi_eos_provider.go
            gnmi_helpers.go
            gnmi_helpers_test.go
            snmp.go
            etc.

it's now:

arista/
    provider/
        eos_provider.go
        gnmi_provider.go
        helpers.go
        provider.go
        cvp/
        darwin/
        gnmi/
            gnmi.go
            helpers.go
            helpers_test.go
        gnmieos/
            gnmi_eos.go
        openconfig/
            helpers.go
        snmp/
            snmp.go

I wanted to do the same with the devices package but there we actually
need a single package to point to where all the devices live so that we
can import it from adapter.go and have all device implementations'
init() methods run. The device implementations are much smaller, though,
and less likely to see namespace collisions, so I guess this is OK.

Change-Id: I3f2abb176ea7e5d9f7f9e829c124a160594b3d65