provider: Add SNMP provider unit tests
I've added some very basic unit tests for the SNMP provider using mocked
data. The mocked data is just an edited (for brevity) version of data
returned by an snmpwalk on a dut. The unit tests just check that the
polling methods (i.e., the methods the provider gives to PollForever to
run periodically) form the correct gNMI SetRequest for the given data.

I've also put the DeviceID and CheckAlive methods on the Snmp provider
itself rather than just in the provider file. This cleans up the code
considerably and allows us to remove global variables.

Change-Id: Ia6f1a32780c2a44fabb9b27e6865a39caa09f62f