provider/snmp: Make SNMP provider ignore inactive interfaces
When a target system has a QSFP interface configured into single-port
mode, it will include all interfaces in responses to SNMP requests for
LLDP system data--including the interfaces that are inactive. For
example, if interface Ethernet1 is QSFP, SNMP requests for interface
data will return only Ethernet1/1, but requests for LLDP data will
return Ethernet1/1, Ethernet1/2, Ethernet1/3, and Ethernet1/4.

This causes the SNMP provider's LLDP updates to fail, since in the
OpenConfig model
/lldp/interfaces/interface[name='Ethernet1/2']/config/name is a leafref
to /oc-if:interfaces/oc-if:interface/oc-if:name.

The fix is to keep track of the interfaces we saw on our last SNMP
ifTable poll and check against that interface list when we get an
lldpPortId PDU. If we didn't see the interface in ifTable, it's inactive
and we should ignore it.

I've added a test case for the bug.

Change-Id: Id9df97368fe4e1aecec9e6687ed063d63cf94c20