provider: Add support for registering Mojo APs via MWM
I've added basic functionality for interacting with MWM and having
Mojo APs register as devices in Aeris.

The manager handles all of the interaction with MWM for now. This might
change eventually if there's some reason to have the Providers make
separate requests. Since everything we need now is contained in the
manageddevices API, though, it makes more sense to make just one
request, unmarshal the JSON into ManagedDevice structs, and then send
just the ManagedDevice struct to the provider. The provider's job is to
form a GNMI SetRequest from the ManagedDevice struct. The Device doesn't
do anything interesting, as it's passed its deviceID on startup and it
always returns true for CheckAlive.

For now auth is pretty weird. We don't have an API key, so the way it
works is that I log into launchpad.wifi.arista.com and get the API
endpoint and JSESSIONID cookie that's used in the CloudVision wifi app.
Those then get fed into the Collector as manageroptions.

I've added a unit test for the Mojo provider and I've manually tested
running a Collector with -manager mwm -manageroption apiurl=<URL>
-manageroption jsessionid=<session ID>. The updates show up in Aeris as
expected.

Change-Id: Ic35e744e9542053d735055cabd0d417ced77f464