provider: refactor gnmi provider
gnmi provider is refactored into helpers.go (opensource), and
client.go (not opensource). Specifically helpers.go contains
Unmarshal(), TestClient struct, various Paths for interface and lldp,
and PollForever. client.go contains the actual implementation of the
gnmi client that will get passed in to GNMIOpenConfigProvider, which
includes a gnmi server of an openconfig tree.

One worth noting point apart from simply moving codes around: The
previous TestPollForever is actually a mix of two tests simultaneously:
we get the SetRequest we want, and we convert the SetRequest into
types.Notification correctly. The latter part is now removed because 1. it's not
applicable because we don't opensource types and anything related to
that, and 2. the convertion is actually already tested by TestSetRequestNotifications
located in the new client_test.go file.

Change-Id: Iafa81be99a13731a35f0b31c1b8b3276e1a23964