provider/snmp: Fix SNMP provider bugs for N9K, PAN devices
We discovered three bugs in the SNMP provider in testing against N9K and
PAN devices:

(1) We can't assume a device serial number will be at
entPhysicalSerialNum.1; instead, we have to walk entPhysicalSerialNum
and take whatever the first serial is. (Actually a better way to do this
is to find the first entPhysicalDescr of type "chassis" and take the
serial number corresponding to that index. But that can wait.)

(2) We can't assume sysName is a fully qualified domain name; sometimes
it'll just be a hostname. In these cases we'll just set the hostname.

(3) OpenConfig doesn't want multi-line strings. So before setting a
string we'll remove newlines and replace them with spaces.

I've added test cases to the unit test and tested manually against the
devices in question.

Change-Id: Icd9866afce600340521f54f2deaf3e472f352c28