provider/snmp: Improve method for selecting serial number
I've changed the way the SNMP provider selects a serial number, so that
rather than selecting the first serial number it sees, it now selects
the first index whose entPhysicalClass is chassis(3), and then takes the
serial number of that index. If no such index exists, we fall back to
taking the first non-empty serial number we see.

I've added unit test cases and tested manually against the N9K and PAN
devices.

Change-Id: I100de80a4e8994e11752d6d98e3f204ca6bed83a