provider/snmp: Add LLDPV2 support
I've added support for the LLDP-V2-MIB to the SNMP provider. The only
thing that's really different is the way the table is indexed, so most
of the code remains the same, except that now we have parallel sets of
OIDs to look for.

I've added a unit test case and tested manually against devices that
support v1 and v2.

Change-Id: I95e4248d466e4f4490520bb4bb7ff8cb5001c9a3