provider/snmp: Clean up SNMP provider
I've done some minor cleanup in the SNMP provider. I've removed a bunch
of variables we don't actually need in the Snmp struct. Each updater
needs to keep track of the indexes it has seen, but those don't need to
be visible to other updaters. I also changed the structure of
updatePlatform to be like the others--there's no longer any reason to
store PDUs in a struct before creating the necessary updates.

I've tested the changes against some physical devices to double-check.

Change-Id: I362616356e0d58c757fbe5e1155f25032cdce5be