Collector: Integration test for supported devices
This commit sets up the infrastructure for testing whether we support a
certain device.

There are two parts of this: one to generate dumps for the devices we
want to do tests on and upload them. Then we have a test that downloads
the dumps and do tests on them.

A script is added that processes configs and runs collector binaries.
To generate a dump for your device, you'll need to put a yaml file in
cmd/Collector/supported_devices/ that specifies the device/manager name
you want to run and the -deviceoption/-manageroption you want.
`make gendumps` will generate dumps locally and `make upload-dumps` will
upload the generated dumps. An example config file is put in the
directory and it's corresponding dump is uploaded.

The integration test relies on turbine vdc testing infrastructure which
gives us the ability to conveniently run any turbines we want. The tests
it does right now is pretty simple as this commit only aims to set up
the infrastructure. The integration test can use both local or remote
dumps for easier manual testing. There isn't a mechanism to pick the
"latest" dump to run tests on for now, which could be added later if
necessary.

closes: https://trello.com/c/2Mevx7no

Change-Id: Icf963febd17ed730cebd3ff5384858089f5b9441