device: Add darwin device, provider to cloudvision-go
I've adapted the existing darwin provider to implement
GNMIOpenConfigProvider and added some functionality (polling for
interface statistics) to flesh it out as an example for the GitHub
repo documentation. Running the device against my laptop now produces a
stream of updates in Aeris.

I've also replaced "XXX_jcr" with "XXX NOTE" in code that's going to be
public.

Change-Id: Id2ac2b53e982aedb52161f77914b1d568e30af96