cmd/Collector: Add Collector binary to cloudvision-go
We've decided to split the device and provider functionality from the
OpenConfig server and Aeris client functionality. The new design is to
have a Collector binary with no plugins (just built-in devices and
providers) that's buildable using only code from cloudvision-go.
Collector will communicate with the Aeris client binary (called Adapter,
CollectorServer, CollectorTranslator, or whatever) over a simple gNMI
connection.

Accordingly, these changes move the Collector main.go into
cloudvision-go, removing the now-unnecessary arguments and adding the
address of the gNMI OpenConfig server (used by GNMIOpenConfigProviders)
and the gNMI pass-through server (used by GNMIPRoviders).

I've gotten rid of the basic manager--the model now is that a manager is
only created when it's specified in the config. Otherwise we just create
the inventory directly from the device config. I've added Yanxi's
device.Config and the skeleton of the code that will accept a config
file, but haven't yet added the code for reading config from a file.

For now the gNMI server addresses are configurable. We may want to
change that in the future, to use a specific pair of ports.

I've tested this code on a dummy gNMI server that just prints out any
SetRequests it receives.

Change-Id: Ic05044a656b8212dcfcbafa708cec874fe520bf2