provider/snmp: Add device-specific SNMP provider unit tests
We want to be able to run unit tests that check the SNMP provider's
output against SNMP responses generated by the provider against a real
target device. This will give us some certainty that a change to the
SNMP provider logic hasn't broken support for specific device types.

I've added the the SNMP provider an option that causes it to generate a
dump of its SNMP interactions with a target device, and another option
that allows the provider to run in a mock mode where it doesn't expect
to make real network requests.

The device unit tests specify a dump file (generated with the former
option) and run the SNMP provider against that dump (using the latter
option).

I've added dumps for an Arista 7150 and 7508, as well as the Cisco and
PAN devices we've tested on in the past.

Change-Id: I1dd055eef4251645a614ab255560d74a742e4249