device: Add typeCheck metadata field
This is to take care of the special case where gnmi provider doesn't
want type checking but still fits openconfig schema. TypeCheck fields is
set and used only internally (can't be set by the user) such that for
gnmi provider openConfig = true but typeCheck = false, for anything else
the two values are the same.

Sanity check by running the dumper against the build.

Change-Id: I207ba2e8cc5b58e0f98dcfc591488bfd4baefd05