provider/snmp: Add support for A10, Cisco, Juniper, Ubiquiti devices
I made a series of changes to support some new device types whose SNMP
responses were breaking the SNMP provider:
- Use lldpLocPortId or lldpLocPortDesc for the LLDP interface
names--whichever matches intfDesc. The way I've done this is just to
check both for a string that matches an intfDesc, and when I find one I
commit to either lldpLocPortId or lldpLocPortDesc from then on.
- For DeviceID, don't just fail if we can't get a serial number. First
try for serial number, then try for lldpLocChassisId, and finally, if
that fails, use the device address.
- for sysname, first try sysName as we've been doing, and then try
lldpLocSysName.

These changes can go in now. I'll add the dumps themselves in separate
commit, and will merge them if/when we find out that's OK.

Change-Id: I7b75c89af67903bada3e95aaf9c402a3b1e4b577