Collector: Switch -mockCheckPath <path>=<feature> to <feature>=<path>
I made a minor tweak in the -mockCheckPath flag. The feature should go
first, I think, since (a) the feature will be short and the path will be
long and (b) the feature is really the thing we care about, and it could
be checked with multiple paths.

For the same reason, I've changed -mockCheckPath to -mockFeature. It's
really the feature we care about, not the path. Also mockFeature is
shorter. :)

Change-Id: I7ab89f29989def28097c52fb7561b7341be2c87f