provider/snmp: Set Snmp.initialized in snmpNetworkInit
The `initialized` bool had been intended to signify that InitGNMI had
been called, and that it was therefore safe to run Snmp.Run(). It has
been tucked into snmpNetworkInit now, though, and now seems to mean
that network initialization is finished. So I've made it that
explicitly, and now we'll just check in Run() that there's a GNMIClient
before starting.

Change-Id: Ic49c7a6f8af1c1c1e24d9a3d3b30db2bf85d0448