provider/snmp: Handle lldpRemPortId being a hex string
The existing code assumes lldpRemPortId values are normal octet strings,
but if the corresponding lldpRemPortIdSubtype is macAddress then the
value may actually be a hex string. We should be doing the same thing
with lldpRemPortId as with lldpRemChassisId--storing the subtype and
possibly interpreting the port ID as a hex string if the subtype is
macAddress.

Change-Id: I6ec05f4f8fc28618cc3fd1c300eb2c52701255c1