provider/snmp: Protect gosnmp network accesses with a lock
We have a lock in the SNMP provider to protect accesses to provider
state, such as interface and LLDP data stored over the course of a poll
cycle. This is actually different from what the lock was originally
intended to do, which was to protect access to the gosnmp connection
object, which isn't set up to handle parallel operations.

Having two locks could result in deadlock if one thread grabs the
provider lock and then needs the connection lock, while another has the
connection lock and then needs the provider lock. We can prevent this by
never having a thread grab the provider lock while the connection lock
is already grabbed. In practice this is easy, since the routines that
handle provider-level state call the routines that use the gosnmp
connection object, and never vice versa.

Alternatively, we could use a single lock to protect all accesses to the
connection state and provider state. This would mean locking at the lvel
of Run(), DeviceID(), Alive(), and the update* methods, and not in the
methods that actually touch the connection object (get(), walk(),
snmpNetworkInit()). To me it's simpler to have those functions always
grab their own lock and never have to worry about the provider lock
(since they needn't access contended provider state), while the methods
that update provider state can worry about their own lock and not have
to worry at all about the connection object.

Change-Id: I6d60196cfc010d6dbb9ca87b7b5f337efd2cdf7a